Add an empty search box and hidden result area
This adds a div with two (empty) selectors and a date input.
Maybe uppon the website loading some JS will populate the locations.
Maybe also upon selecting an origin the destination will change.
Or we could add a graph traversal thing in JS and offer trips with
connections.
Anyway.
There is a div that is not displayed beneath that, which could hold
a form filled with options related to possible journeys plus a
submit button, which could lead to the actuall "buy a ticket" page.