Remove an added span in form
There was a span I added in the form that would show currently selected
locations, but I appended it after the area that gets cleaned when
the user makes a search, so it would get appended again and again after
every search.
I could move it somewhere else, or change the layout of that container,
but since it is not important to show this at that place again, I am
just removing this for now.