Make popup background unclickable
This makes the inner div work not only as a mere grid container but also
as the actual pop up (I have given it the css attributes re: margins
and corners etc that were originally part of the outer div).
Moreover, the shadow of the popup is now the background of the outer div
(and is also transparent-ish).