Make the graph more standard and include IDs
The edge list used to have variable sized lists for each connection and
now it has pairs of node ids to denote an edge.
Moreover, I didn't want to have to change the edge list if I ever made
a change in the node list (the indices would change) so I made the node
list keep names and ids and started writing this as a value in the html
tags,.
The downside to this is that once I had gotten the index of the name of
the chosen origin, writing the destination name was a constant-time
operation and now for each destination I have to loop through the nodes
and check the ids.