transform announcement list to handlebar format
This commit changes the way the announcement list in the sidebar is
populated. Instead of it being hard-coded in the clientside js file,
it is rendered serverside.
The only problem is that I have not yet created the appropriate database
support for it, so the information remain hardcoded (in a js file in the
root directory instead)