rewrite of the trip planner
In this commit the client-side js responsible for filling out the two
location selectors is simplified. The server is now responsible for
filling out both (keeping all destinations hidden) while also encoding
the direct connections in a div struct at the bottom of the page (this
feels like a hack) and the client is responsible for enabling (and doing
this appropriate resets) the valid destinations based on user input.
Once again, the information for this part is written outside of the sql
database (pending a re-write to facilitate the connection).
Moreover, the departures part of the original page has been disabled
since I think it will have to be changed significantly when the db
connection is finally established.